@@ -1,95 +1,54 @@
 {
-  "name": "angular",
-  "version": "1.0.0",
-  "license": "MIT",
-  "author": {
-    "name": "wangdong",
-    "email": "mail@wangdong.io",
-    "url": "https://wangdong.io"
-  },
-  "engines": {
-    "node": ">=8.0.0"
-  },
-  "main": "server",
-  "private": true,
+  "name": "material",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "pm2": "pm2",
-    "watch": "npm run make && ng serve",
-    "build": "npm run make && ng build --prod",
+    "parse": "parse.server",
+    "start": "ng serve",
+    "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "make": "node script/make.js",
-    "setup": "node script/setup.js",
-    "backup": "node script/backup.js",
-    "restore": "node script/restore.js",
-    "start": "pm2 start ecosystem.json",
-    "restart": "pm2 restart ecosystem.json",
-    "reload": "pm2 reload ecosystem.json",
-    "status": "pm2 status ecosystem.json",
-    "delete": "pm2 delete ecosystem.json",
-    "logs": "pm2 logs",
-    "monit": "pm2 monit",
-    "docker": "docker-compose exec server npm run"
+    "e2e": "ng e2e"
   },
+  "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.0",
-    "@angular/cdk": "^5.2.4",
-    "@angular/common": "^5.2.0",
-    "@angular/compiler": "^5.2.0",
-    "@angular/core": "^5.2.0",
-    "@angular/flex-layout": "5.0.0-beta.14",
-    "@angular/forms": "^5.2.0",
-    "@angular/http": "^5.2.0",
-    "@angular/material": "^5.2.4",
-    "@angular/platform-browser": "^5.2.0",
-    "@angular/platform-browser-dynamic": "^5.2.0",
-    "@angular/router": "^5.2.0",
-    "args": "^4.0.0",
-    "core-js": "^2.4.1",
-    "dotenv": "^5.0.1",
-    "express": "^4.16.3",
-    "font-awesome": "^4.7.0",
+    "@angular/animations": "~8.0.0",
+    "@angular/cdk": "^8.0.1",
+    "@angular/common": "~8.0.0",
+    "@angular/compiler": "~8.0.0",
+    "@angular/core": "~8.0.0",
+    "@angular/flex-layout": "~8.0.0-beta.26",
+    "@angular/forms": "~8.0.0",
+    "@angular/material": "^8.0.1",
+    "@angular/platform-browser": "~8.0.0",
+    "@angular/platform-browser-dynamic": "~8.0.0",
+    "@angular/router": "~8.0.0",
+    "@wang-dong/parse.server": "^1.0.0",
     "hammerjs": "^2.0.8",
-    "highlight.js": "^9.12.0",
-    "lg-autoplay.js": "^1.0.0",
-    "lg-fullscreen.js": "^1.0.0",
-    "lg-thumbnail.js": "^1.0.0",
-    "lg-zoom.js": "^1.0.1",
-    "lightgallery.js": "^1.0.2",
-    "moment": "^2.22.0",
-    "mongodb-backup": "^1.6.9",
-    "mongodb-restore": "^1.6.2",
-    "nodemailer": "^4.6.4",
-    "parse-dashboard": "^1.2.0",
-    "parse-server": "^2.7.4",
-    "perfect-scrollbar": "^1.3.0",
-    "pm2": "^2.10.2",
-    "qrious": "^4.0.2",
-    "rxjs": "^5.5.6",
-    "sendmail": "^1.4.1",
-    "simplemde": "^1.11.2",
-    "zone.js": "^0.8.19"
+    "material-design-icons": "^3.0.1",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "~1.7.4",
-    "@angular/compiler-cli": "^5.2.0",
-    "@angular/language-service": "^5.2.0",
-    "@types/jasmine": "~2.8.3",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
-    "jasmine-core": "~2.8.0",
+    "@angular-devkit/build-angular": "^0.800.2",
+    "@angular/cli": "^8.0.2",
+    "@angular/compiler-cli": "~8.0.0",
+    "@angular/language-service": "~8.0.0",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "^5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~2.0.0",
+    "karma": "~4.1.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~4.1.0",
-    "tslint": "~5.9.1",
-    "typescript": "~2.5.3"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.4.3"
   }
 }
