@@ -1,15 +1,16 @@
-import {NgModule} from '@angular/core';
-import {SharedModule} from '~shared/shared.module';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { MatButtonModule } from '@angular/material';
 
-import {ErrorRoutingModule} from './error-routing.module';
-import {ErrorComponent} from './error.component';
+import { ErrorRoutingModule } from './error-routing.module';
+import { ErrorComponent } from './error.component';
 
 @NgModule({
+  declarations: [ErrorComponent],
   imports: [
-    SharedModule,
+    CommonModule,
+    MatButtonModule,
     ErrorRoutingModule
-  ],
-  declarations: [ErrorComponent]
+  ]
 })
-export class ErrorModule {
-}
+export class ErrorModule { }
