@@ -1,16 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 
-import { LoginComponent } from './login.component';
+import { IndexComponent } from './index.component';
 
 const routes: Routes = [
   {
-    path: 'login',
-    component: LoginComponent,
-    data: {
-      fullscreen: true,
-      auth: false
-    }
+    path: '',
+    component: IndexComponent
   }
 ];
 
@@ -18,4 +14,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class LoginRoutingModule { }
+export class IndexRoutingModule { }
