@@ -1,17 +1,35 @@
-import {NgModule} from '@angular/core';
-import {SharedModule} from '~shared/shared.module';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import {
+  MatCardModule,
+  MatFormFieldModule,
+  MatInputModule,
+  MatButtonModule,
+  MatIconModule,
+  MatSnackBarModule
+} from '@angular/material';
 
-import {LoginRoutingModule} from './login-routing.module';
-import {LoginComponent} from './login.component';
-import {UserService} from './user.service';
+import { LoginRoutingModule } from './login-routing.module';
+import { LoginComponent } from './login.component';
 
 @NgModule({
-  imports: [
-    SharedModule,
-    LoginRoutingModule
+  declarations: [
+    LoginComponent
   ],
-  declarations: [LoginComponent],
-  providers: [UserService]
+  imports: [
+    CommonModule,
+    FormsModule,
+    ReactiveFormsModule,
+    FlexLayoutModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatButtonModule,
+    MatIconModule,
+    MatSnackBarModule,
+    LoginRoutingModule
+  ]
 })
-export class LoginModule {
-}
+export class LoginModule { }
