@@ -1,3 +1,8 @@
 export const environment = {
-  production: true
+  production: true,
+  parse: {
+    appId: 'A9bVJTT5fmH5zqP8CPqDK1cva5NfkBP6',
+    javascriptKey: '96SFBryScCCydhqwjG8ZhnsKjUC1jB5g',
+    serverURL: 'http://127.0.0.1:8080/parse'
+  }
 };
